requirements.txt
# Project dependencies
openai
colorama

pandas
numpy
markdown
beautifulsoup4

dspy-ai